[del] VolumeticMesh.cs
